Move speedy compiler in is own package
