+++ sdk/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/package.scala
@@ -1,0 +1,11 @@
+// Copyright (c) 2025 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.digitalasset.daml.lf
+
+package object speedy {
+
+  val Compiler = compiler.Compiler
+  type Compiler = compiler.Compiler
+
+}
