+++ sdk/daml-lf/interpreter/src/test/scala/com/digitalasset/daml/lf/speedy/compiler/PhaseOneTest.scala
@@ -1,16 +1,18 @@
 // Copyright (c) 2025 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.daml.lf.speedy
+package com.digitalasset.daml.lf
+package speedy
+package compiler
 
 import com.digitalasset.daml.lf.data.ImmArray
 import com.digitalasset.daml.lf.data.Ref._
 import com.digitalasset.daml.lf.data.Ref.PackageId
 import com.digitalasset.daml.lf.language.Ast._
 import com.digitalasset.daml.lf.language.PackageInterface
-import com.digitalasset.daml.lf.speedy.ClosureConversion.closureConvert
-import com.digitalasset.daml.lf.speedy.SExpr0._
-import com.digitalasset.daml.lf.speedy.Anf.flattenToAnf
+import com.digitalasset.daml.lf.speedy.compiler.ClosureConversion.closureConvert
+import com.digitalasset.daml.lf.speedy.compiler.SExpr0._
+import com.digitalasset.daml.lf.speedy.compiler.Anf.flattenToAnf
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.prop.TableDrivenPropertyChecks
